requirements.txt
boto3>=1.26.0
requests>=2.28.0
pyyaml>=6.0
tabulate>=0.9.0
pytest>=7.0.0
python-dotenv>=1.0.0
PyGithub>=1.58.0 